requirements.txt
google-generativeai==0.4.1
simple-term-menu==1.6.4
prompt_toolkit==3.0.38
pyperclip==1.8.2
rich==13.3.2
